requirements.txt
SQLAlchemy==0.9.3
path.py==5.1
ipython==1.2.0
alembic==0.6.3
DateTime==4.0.1
Flask==0.10.1
Flask-SQLAlchemy==1.0
Flask-Security==1.7.1
py-bcrypt==0.4
pudb==2013.5.1
Flask-Babel==0.9
Flask-WTF==0.9.4
celery[librabbitmq,redis,auth,msgpack]==3.1.9
python-memcached==1.53
MySQL-python==1.2.5
gunicorn==18.0
Flask-Testing==0.4
nose==1.3.0